notes_project/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
pytest 

